chore: multiple quantity with sys decimals

+++ programs/pie/src/constant.rs
@@ -4,7 +4,7 @@ pub const USER_FUND: &[u8] = b"user_fund";
 pub const BASKET_MINT: &[u8] = b"basket_mint";
 pub const BASKET_CONFIG: &[u8] = b"basket_config";
 
-pub const SYS_DECIMALS: u128 = 1_000_000;
+pub const SYS_DECIMALS: u64 = 1_000_000;
 pub const MAX_COMPONENTS: u8 = 50;
 
 pub const INITIALIZE_ADMIN: &str = "DjCDT99HEZyZuq48wugooYNGBnrB5Nhe93VgAKKBAPDV";
